all new updates
all

+++ src/streams/comparator/comparators/EmployeeSalaryComparator.java
@@ -13,6 +13,6 @@ public class EmployeeSalaryComparator implements Comparator<Employee> {
  
     @Override
     public int compare(Employee emp1, Employee emp2) {
-        return emp1.getSalary() - emp2.getSalary();
+        return (int) (emp1.getSalary() - emp2.getSalary());
     }
 }
